[BE] refactor: STATUS 상수로 별도 분리 #42
