+++ server/src/service/minos/L.mino.ts
@@ -1,6 +1,6 @@
 import Mino from '@/service/minos/mino';
 import Position from '@/types/position';
-import STATUS from '@/service/status';
+import STATUS from '@/service/constants';
 
 class LMino extends Mino {
   color: string;
